pyproject.toml
[project]
name = "FinancialCareerCommunity"
version = "0.1.0"
description = "A web application for managing a financial career community."
authors = [
    { name = "Muika(@Moemu)", email = "i@snowy.moe" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colorlog>=6.9.0",
    "pre-commit>=4.3.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
